changes to psp_malloc

+++ src/utils/psp_malloc.h
@@ -3,6 +3,6 @@
 
 void init_malloc();
 
-void * psp_malloc(size_t bytes, const char * debug_info = "");
+void * psp_malloc(unsigned int bytes, const char * debug_info = "");
 
 void psp_free(void * ptr);
